Update endorsement.md
If a work is in an academic journal, it's been endorsed.
If a work has been referenced in a conference like IEEE, it's been endorsed. 

We need to let endorsement me a technical gate keep, but not something that is transparently censor works and not allowing them to be taking up by databases. 

If you guys need more time to look at that for my own, that's fine. But it's been endorsed. I would like to continue with the DevsOps work now please. Not the semantics. 

Want to put me on Reddit? K, we'll I'm hoping we're better than that for our College. This is a bit much when it comes to gatekeeping technically speaking.